Refactor: Update auth import paths and toast variants
