+++ app/api/admin/delivery-zones/[id]/route.ts
@@ -6,7 +6,7 @@
 
 import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
-import { auth } from '@/server/auth'
+import { auth } from '@/lib/auth'
 
 export async function PATCH(
   req: NextRequest,
