+++ app/api/admin/orders/route.ts
@@ -5,7 +5,7 @@
 
 import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
-import { auth } from '@/server/auth'
+import { auth } from '@/lib/auth'
 
 export async function GET(req: NextRequest) {
   try {
